16-th homework (morphs, debugbar, policy) - complete

+++ app/Http/Controllers/Admin/AdminCategoryController.php
@@ -1,13 +1,15 @@
 <?php
 
-namespace App\Http\Controllers;
+namespace App\Http\Controllers\Admin;
 
+use App\Http\Controllers\Controller;
 use App\Models\Category;
 use Illuminate\Http\Request;
 
 
-class AdminCategoryController
+class AdminCategoryController extends Controller
 {
+
     public function index()
     {
         $categories = Category::paginate(10);
@@ -16,6 +18,7 @@ class AdminCategoryController
 
     public function create()
     {
+        $this->authorize('create', Category::class);
         $category = new Category();
         return view('admin/category/form', compact('category'));
     }
@@ -33,6 +36,7 @@ class AdminCategoryController
 
     public function edit($id)
     {
+        $this->authorize('create', Category::class);
         $category = Category::find($id);
         return view('admin/category/form-edit', compact('category'));
     }
@@ -50,6 +54,7 @@ class AdminCategoryController
 
     public function destroy(Request $request)
     {
+        $this->authorize('create', Category::class);
         $category = Category::find($request->id);
         $category->delete();
         return redirect()->route('admin.category');
